Split web PMA route and managed-thread harness into focused support layers
Extract repeated inline fake classes from test_pma_managed_threads_messages.py
into tests/pma_support/managed_threads.py, reducing the test file by 568 lines
while keeping all 26 tests passing.
