+++ tests/pma_support/core.py
@@ -1,9 +1,5 @@
 import asyncio
-import concurrent.futures
 import json
-import os
-import threading
-from datetime import datetime, timezone
 from pathlib import Path
 from types import SimpleNamespace
 from typing import Any, Optional
@@ -15,15 +11,12 @@ from fastapi.testclient import TestClient
 
 from codex_autorunner.agents.opencode.runtime import OpenCodeTurnOutput
 from codex_autorunner.agents.registry import AgentDescriptor
-from codex_autorunner.bootstrap import pma_active_context_content, seed_hub_files
-from codex_autorunner.core import filebox
 from codex_autorunner.core.config import CONFIG_FILENAME, DEFAULT_HUB_CONFIG
 from codex_autorunner.core.orchestration import (
     ColdTraceStore,
     ExecutionRecord,
     ThreadTarget,
 )
-from codex_autorunner.core.pma_context import maybe_auto_prune_active_context
 from codex_autorunner.core.pma_queue import PmaQueue, QueueItemState
 from codex_autorunner.core.pma_thread_store import PmaThreadStore
 from codex_autorunner.core.pma_transcripts import PmaTranscriptStore
@@ -44,133 +37,15 @@ from codex_autorunner.surfaces.web.routes.pma_routes import (
 )
 from codex_autorunner.surfaces.web.routes.pma_routes import publish as publish_routes
 from tests.conftest import write_test_config
+from tests.pma_support import (
+    _disable_pma,
+    _enable_pma,
+    _install_fake_successful_chat_supervisor,
+    _seed_discord_pma_binding,
+    _seed_telegram_pma_binding,
+)
 
 pytestmark = pytest.mark.slow
-
-
-def _enable_pma(
-    hub_root: Path,
-    *,
-    model: Optional[str] = None,
-    reasoning: Optional[str] = None,
-    max_text_chars: Optional[int] = None,
-) -> None:
-    cfg = json.loads(json.dumps(DEFAULT_HUB_CONFIG))
-    cfg.setdefault("pma", {})
-    cfg["pma"]["enabled"] = True
-    if model is not None:
-        cfg["pma"]["model"] = model
-    if reasoning is not None:
-        cfg["pma"]["reasoning"] = reasoning
-    if max_text_chars is not None:
-        cfg["pma"]["max_text_chars"] = max_text_chars
-    write_test_config(hub_root / CONFIG_FILENAME, cfg)
-
-
-def _disable_pma(hub_root: Path) -> None:
-    cfg = json.loads(json.dumps(DEFAULT_HUB_CONFIG))
-    cfg.setdefault("pma", {})
-    cfg["pma"]["enabled"] = False
-    write_test_config(hub_root / CONFIG_FILENAME, cfg)
-
-
-def _install_fake_successful_chat_supervisor(
-    app,
-    *,
-    turn_id: str,
-    message: str = "assistant text",
-    raw_events: Optional[list[dict]] = None,
-) -> None:
-    class FakeTurnHandle:
-        def __init__(self) -> None:
-            self.turn_id = turn_id
-
-        async def wait(self, timeout=None):
-            _ = timeout
-            return type(
-                "Result",
-                (),
-                {
-                    "agent_messages": [message],
-                    "raw_events": list(raw_events or []),
-                    "errors": [],
-                },
-            )()
-
-    class FakeClient:
-        async def thread_resume(self, thread_id: str) -> None:
-            _ = thread_id
-            return None
-
-        async def thread_start(self, root: str) -> dict:
-            _ = root
-            return {"id": "thread-1"}
-
-        async def turn_start(
-            self,
-            thread_id: str,
-            prompt: str,
-            approval_policy: str,
-            sandbox_policy: str,
-            **turn_kwargs,
-        ):
-            _ = thread_id, prompt, approval_policy, sandbox_policy, turn_kwargs
-            return FakeTurnHandle()
-
-    class FakeSupervisor:
-        def __init__(self) -> None:
-            self.client = FakeClient()
-
-        async def get_client(self, hub_root: Path):
-            _ = hub_root
-            return self.client
-
-    app.state.app_server_supervisor = FakeSupervisor()
-
-
-async def _seed_discord_pma_binding(hub_env, *, channel_id: str) -> None:
-    store = DiscordStateStore(
-        hub_env.hub_root / ".codex-autorunner" / "discord_state.sqlite3"
-    )
-    try:
-        await store.upsert_binding(
-            channel_id=channel_id,
-            guild_id="guild-1",
-            workspace_path=str(hub_env.repo_root.resolve()),
-            repo_id=hub_env.repo_id,
-        )
-        await store.update_pma_state(
-            channel_id=channel_id,
-            pma_enabled=True,
-            pma_prev_workspace_path=str(hub_env.repo_root.resolve()),
-            pma_prev_repo_id=hub_env.repo_id,
-        )
-    finally:
-        await store.close()
-
-
-async def _seed_telegram_pma_binding(
-    hub_env, *, chat_id: int, thread_id: Optional[int]
-) -> None:
-    store = TelegramStateStore(
-        hub_env.hub_root / ".codex-autorunner" / "telegram_state.sqlite3"
-    )
-    try:
-        key = topic_key(chat_id, thread_id)
-        await store.bind_topic(
-            key,
-            str(hub_env.repo_root.resolve()),
-            repo_id=hub_env.repo_id,
-        )
-
-        def _enable(record: Any) -> None:
-            record.pma_enabled = True
-            record.pma_prev_repo_id = hub_env.repo_id
-            record.pma_prev_workspace_path = str(hub_env.repo_root.resolve())
-
-        await store.update_topic(key, _enable)
-    finally:
-        await store.close()
 
 
 def test_build_pma_routes_does_not_construct_async_primitives_on_route_build(
@@ -2586,1260 +2461,6 @@ def test_pma_managed_thread_status_and_tail_use_orchestration_service(
     ]
     assert ("get_queue_depth", "thread-1") in fake_service.calls
     assert fake_service.calls.count(("get_running_execution", "thread-1")) >= 3
-
-
-def test_pma_files_list_empty(hub_env) -> None:
-    seed_hub_files(hub_env.hub_root, force=True)
-    _enable_pma(hub_env.hub_root)
-    app = create_hub_app(hub_env.hub_root)
-    client = TestClient(app)
-
-    resp = client.get("/hub/pma/files")
-    assert resp.status_code == 200
-    payload = resp.json()
-    assert payload["inbox"] == []
-    assert payload["outbox"] == []
-
-
-def test_pma_files_upload_list_download_delete(hub_env) -> None:
-    seed_hub_files(hub_env.hub_root, force=True)
-    _enable_pma(hub_env.hub_root)
-    app = create_hub_app(hub_env.hub_root)
-    client = TestClient(app)
-
-    # Upload a file to inbox
-    files = {"file.txt": ("file.txt", b"Hello, PMA!", "text/plain")}
-    resp = client.post("/hub/pma/files/inbox", files=files)
-    assert resp.status_code == 200
-
-    # List files
-    resp = client.get("/hub/pma/files")
-    assert resp.status_code == 200
-    payload = resp.json()
-    assert len(payload["inbox"]) == 1
-    assert payload["inbox"][0]["name"] == "file.txt"
-    assert payload["inbox"][0]["box"] == "inbox"
-    assert payload["inbox"][0]["size"] == 11
-    assert payload["inbox"][0]["source"] == "filebox"
-    assert payload["inbox"][0]["next_action"] == "process_uploaded_file"
-    assert payload["inbox"][0]["likely_false_positive"] is False
-    assert "/hub/pma/files/inbox/file.txt" in payload["inbox"][0]["url"]
-    assert payload["outbox"] == []
-    assert (
-        filebox.inbox_dir(hub_env.hub_root) / "file.txt"
-    ).read_bytes() == b"Hello, PMA!"
-
-    # Download file
-    resp = client.get("/hub/pma/files/inbox/file.txt")
-    assert resp.status_code == 200
-    assert resp.content == b"Hello, PMA!"
-
-    # Delete file
-    resp = client.delete("/hub/pma/files/inbox/file.txt")
-    assert resp.status_code == 200
-
-    # Verify file is gone
-    resp = client.get("/hub/pma/files")
-    assert resp.status_code == 200
-    payload = resp.json()
-    assert payload["inbox"] == []
-
-
-def test_pma_files_invalid_box(hub_env) -> None:
-    seed_hub_files(hub_env.hub_root, force=True)
-    _enable_pma(hub_env.hub_root)
-    app = create_hub_app(hub_env.hub_root)
-    client = TestClient(app)
-
-    # Try to upload to invalid box
-    files = {"file.txt": ("file.txt", b"test", "text/plain")}
-    resp = client.post("/hub/pma/files/invalid", files=files)
-    assert resp.status_code == 400
-
-    # Try to download from invalid box
-    resp = client.get("/hub/pma/files/invalid/file.txt")
-    assert resp.status_code == 400
-
-    # Try to delete from invalid box
-    resp = client.delete("/hub/pma/files/invalid/file.txt")
-    assert resp.status_code == 400
-
-
-def test_pma_files_list_ignores_legacy_sources(hub_env) -> None:
-    seed_hub_files(hub_env.hub_root, force=True)
-    _enable_pma(hub_env.hub_root)
-    app = create_hub_app(hub_env.hub_root)
-    client = TestClient(app)
-
-    filebox.ensure_structure(hub_env.hub_root)
-    (filebox.inbox_dir(hub_env.hub_root) / "primary.txt").write_bytes(b"primary")
-    legacy_pma = hub_env.hub_root / ".codex-autorunner" / "pma" / "inbox"
-    legacy_pma.mkdir(parents=True, exist_ok=True)
-    (legacy_pma / "legacy-pma.txt").write_bytes(b"legacy-pma")
-    legacy_telegram = (
-        hub_env.hub_root
-        / ".codex-autorunner"
-        / "uploads"
-        / "telegram-files"
-        / "topic-1"
-        / "inbox"
-    )
-    legacy_telegram.mkdir(parents=True, exist_ok=True)
-    (legacy_telegram / "legacy-telegram.txt").write_bytes(b"legacy-telegram")
-
-    resp = client.get("/hub/pma/files")
-    assert resp.status_code == 200
-    payload = resp.json()
-    assert len(payload["inbox"]) == 1
-    item = payload["inbox"][0]
-    assert item["box"] == "inbox"
-    assert item["item_type"] == "pma_file"
-    assert item["name"] == "primary.txt"
-    assert item["next_action"] == "process_uploaded_file"
-    assert item["likely_false_positive"] is False
-    assert item["size"] == 7
-    assert item["source"] == "filebox"
-    assert item["url"].endswith("/hub/pma/files/inbox/primary.txt")
-    assert item["modified_at"]
-    assert (item.get("freshness") or {}).get("is_stale") is False
-
-
-def test_pma_files_list_marks_old_inbox_files_as_likely_stale(hub_env) -> None:
-    seed_hub_files(hub_env.hub_root, force=True)
-    _enable_pma(hub_env.hub_root)
-    app = create_hub_app(hub_env.hub_root)
-    client = TestClient(app)
-
-    inbox_file = filebox.inbox_dir(hub_env.hub_root) / "forgotten.txt"
-    inbox_file.parent.mkdir(parents=True, exist_ok=True)
-    inbox_file.write_bytes(b"old")
-    old_ts = datetime(2026, 3, 16, 9, 0, tzinfo=timezone.utc).timestamp()
-    os.utime(inbox_file, (old_ts, old_ts))
-
-    resp = client.get("/hub/pma/files")
-    assert resp.status_code == 200
-    payload = resp.json()
-    assert len(payload["inbox"]) == 1
-    item = payload["inbox"][0]
-    assert item["name"] == "forgotten.txt"
-    assert item["next_action"] == "review_stale_uploaded_file"
-    assert item["likely_false_positive"] is True
-    assert "already handled" in (item["attention_summary"] or "").lower()
-    assert (item.get("freshness") or {}).get("is_stale") is True
-
-
-def test_pma_files_download_rejects_legacy_path(hub_env) -> None:
-    seed_hub_files(hub_env.hub_root, force=True)
-    _enable_pma(hub_env.hub_root)
-    app = create_hub_app(hub_env.hub_root)
-    client = TestClient(app)
-
-    legacy_pma = hub_env.hub_root / ".codex-autorunner" / "pma" / "inbox"
-    legacy_pma.mkdir(parents=True, exist_ok=True)
-    (legacy_pma / "legacy.txt").write_bytes(b"legacy")
-
-    resp = client.get("/hub/pma/files/inbox/legacy.txt")
-    assert resp.status_code == 404
-
-
-def test_pma_files_outbox(hub_env) -> None:
-    seed_hub_files(hub_env.hub_root, force=True)
-    _enable_pma(hub_env.hub_root)
-    app = create_hub_app(hub_env.hub_root)
-    client = TestClient(app)
-
-    # Upload a file to outbox
-    files = {"output.txt": ("output.txt", b"Output content", "text/plain")}
-    resp = client.post("/hub/pma/files/outbox", files=files)
-    assert resp.status_code == 200
-
-    # List files
-    resp = client.get("/hub/pma/files")
-    assert resp.status_code == 200
-    payload = resp.json()
-    assert len(payload["outbox"]) == 1
-    assert payload["outbox"][0]["name"] == "output.txt"
-    assert payload["outbox"][0]["box"] == "outbox"
-    assert "/hub/pma/files/outbox/output.txt" in payload["outbox"][0]["url"]
-    assert payload["inbox"] == []
-
-    # Download from outbox
-    resp = client.get("/hub/pma/files/outbox/output.txt")
-    assert resp.status_code == 200
-    assert resp.content == b"Output content"
-
-
-def test_pma_files_delete_removes_only_resolved_file(hub_env) -> None:
-    seed_hub_files(hub_env.hub_root, force=True)
-    _enable_pma(hub_env.hub_root)
-    app = create_hub_app(hub_env.hub_root)
-    client = TestClient(app)
-
-    filebox.ensure_structure(hub_env.hub_root)
-    (filebox.inbox_dir(hub_env.hub_root) / "shared.txt").write_bytes(b"primary")
-    legacy_pma = hub_env.hub_root / ".codex-autorunner" / "pma" / "inbox"
-    legacy_pma.mkdir(parents=True, exist_ok=True)
-    (legacy_pma / "shared.txt").write_bytes(b"legacy-pma")
-    legacy_telegram = (
-        hub_env.hub_root
-        / ".codex-autorunner"
-        / "uploads"
-        / "telegram-files"
-        / "topic-2"
-        / "inbox"
-    )
-    legacy_telegram.mkdir(parents=True, exist_ok=True)
-    (legacy_telegram / "shared.txt").write_bytes(b"legacy-telegram")
-
-    resp = client.delete("/hub/pma/files/inbox/shared.txt")
-    assert resp.status_code == 200
-    assert resp.json() == {"status": "ok"}
-    assert not (filebox.inbox_dir(hub_env.hub_root) / "shared.txt").exists()
-    assert (legacy_pma / "shared.txt").exists()
-    assert (legacy_telegram / "shared.txt").exists()
-
-
-def test_pma_files_bulk_delete_removes_only_canonical_entries(hub_env) -> None:
-    seed_hub_files(hub_env.hub_root, force=True)
-    _enable_pma(hub_env.hub_root)
-    app = create_hub_app(hub_env.hub_root)
-    client = TestClient(app)
-
-    filebox.ensure_structure(hub_env.hub_root)
-    (filebox.outbox_dir(hub_env.hub_root) / "a.txt").write_bytes(b"a")
-    legacy_pma = hub_env.hub_root / ".codex-autorunner" / "pma" / "outbox"
-    legacy_pma.mkdir(parents=True, exist_ok=True)
-    (legacy_pma / "b.txt").write_bytes(b"b")
-    legacy_telegram_pending = (
-        hub_env.hub_root
-        / ".codex-autorunner"
-        / "uploads"
-        / "telegram-files"
-        / "topic-3"
-        / "outbox"
-        / "pending"
-    )
-    legacy_telegram_pending.mkdir(parents=True, exist_ok=True)
-    (legacy_telegram_pending / "c.txt").write_bytes(b"c")
-
-    resp = client.delete("/hub/pma/files/outbox")
-    assert resp.status_code == 200
-    assert resp.json() == {"status": "ok"}
-    assert (client.get("/hub/pma/files").json()["outbox"]) == []
-    assert not (filebox.outbox_dir(hub_env.hub_root) / "a.txt").exists()
-    assert (legacy_pma / "b.txt").exists()
-    assert (legacy_telegram_pending / "c.txt").exists()
-
-
-def test_pma_files_bulk_delete_leaves_legacy_duplicates_hidden(hub_env) -> None:
-    seed_hub_files(hub_env.hub_root, force=True)
-    _enable_pma(hub_env.hub_root)
-    app = create_hub_app(hub_env.hub_root)
-    client = TestClient(app)
-
-    filebox.ensure_structure(hub_env.hub_root)
-    (filebox.outbox_dir(hub_env.hub_root) / "shared.txt").write_bytes(b"primary")
-    legacy_pma = hub_env.hub_root / ".codex-autorunner" / "pma" / "outbox"
-    legacy_pma.mkdir(parents=True, exist_ok=True)
-    (legacy_pma / "shared.txt").write_bytes(b"legacy")
-
-    resp = client.delete("/hub/pma/files/outbox")
-    assert resp.status_code == 200
-    assert resp.json() == {"status": "ok"}
-    assert not (filebox.outbox_dir(hub_env.hub_root) / "shared.txt").exists()
-    assert (legacy_pma / "shared.txt").exists()
-    payload = client.get("/hub/pma/files").json()
-    assert payload["outbox"] == []
-
-
-def test_pma_files_list_waits_for_bulk_delete_to_finish(hub_env, monkeypatch) -> None:
-    seed_hub_files(hub_env.hub_root, force=True)
-    _enable_pma(hub_env.hub_root)
-    app = create_hub_app(hub_env.hub_root)
-    client = TestClient(app)
-
-    filebox.ensure_structure(hub_env.hub_root)
-    race_file = filebox.inbox_dir(hub_env.hub_root) / "race.txt"
-    race_file.write_bytes(b"race")
-
-    unlink_started = threading.Event()
-    release_unlink = threading.Event()
-    original_unlink = Path.unlink
-
-    def _gated_unlink(path: Path, *args: Any, **kwargs: Any) -> None:
-        if path == race_file and not unlink_started.is_set():
-            unlink_started.set()
-            if not release_unlink.wait(timeout=3):
-                raise AssertionError("Timed out waiting to release bulk delete unlink")
-        original_unlink(path, *args, **kwargs)
-
-    monkeypatch.setattr(Path, "unlink", _gated_unlink)
-
-    with concurrent.futures.ThreadPoolExecutor(max_workers=2) as executor:
-        delete_future = executor.submit(lambda: client.delete("/hub/pma/files/inbox"))
-        assert unlink_started.wait(timeout=3), "Bulk delete did not reach unlink gate"
-        list_future = executor.submit(lambda: client.get("/hub/pma/files"))
-        release_unlink.set()
-
-        delete_resp = delete_future.result(timeout=3)
-        list_resp = list_future.result(timeout=3)
-
-    assert delete_resp.status_code == 200
-    assert list_resp.status_code == 200
-    assert list_resp.json()["inbox"] == []
-
-
-def test_pma_files_rejects_invalid_filenames(hub_env) -> None:
-    seed_hub_files(hub_env.hub_root, force=True)
-    _enable_pma(hub_env.hub_root)
-    app = create_hub_app(hub_env.hub_root)
-    client = TestClient(app)
-
-    # Test traversal attempts - upload rejects invalid filenames
-    for filename in ["../x", "..", "a/b", "a\\b", ".", ""]:
-        files = {"file": (filename, b"test", "text/plain")}
-        resp = client.post("/hub/pma/files/inbox", files=files)
-        assert resp.status_code == 400, f"Should reject filename: {filename}"
-        assert "filename" in resp.json()["detail"].lower()
-
-
-def test_pma_files_size_limit(hub_env) -> None:
-    seed_hub_files(hub_env.hub_root, force=True)
-    _enable_pma(hub_env.hub_root)
-    app = create_hub_app(hub_env.hub_root)
-    client = TestClient(app)
-
-    max_upload_bytes = DEFAULT_HUB_CONFIG["pma"]["max_upload_bytes"]
-
-    # Upload a file that exceeds the size limit
-    large_content = b"x" * (max_upload_bytes + 1)
-    files = {"large.bin": ("large.bin", large_content, "application/octet-stream")}
-    resp = client.post("/hub/pma/files/inbox", files=files)
-    assert resp.status_code == 400
-    assert "too large" in resp.json()["detail"].lower()
-
-    # Upload a file that is exactly at the limit
-    limit_content = b"y" * max_upload_bytes
-    files = {"limit.bin": ("limit.bin", limit_content, "application/octet-stream")}
-    resp = client.post("/hub/pma/files/inbox", files=files)
-    assert resp.status_code == 200
-    assert "limit.bin" in resp.json()["saved"]
-
-
-def test_pma_files_returns_404_for_nonexistent_files(hub_env) -> None:
-    seed_hub_files(hub_env.hub_root, force=True)
-    _enable_pma(hub_env.hub_root)
-    app = create_hub_app(hub_env.hub_root)
-    client = TestClient(app)
-
-    # Download non-existent file
-    resp = client.get("/hub/pma/files/inbox/nonexistent.txt")
-    assert resp.status_code == 404
-    assert "File not found" in resp.json()["detail"]
-
-    # Delete non-existent file
-    resp = client.delete("/hub/pma/files/inbox/nonexistent.txt")
-    assert resp.status_code == 404
-    assert "File not found" in resp.json()["detail"]
-
-
-def test_pma_docs_list(hub_env) -> None:
-    seed_hub_files(hub_env.hub_root, force=True)
-    _enable_pma(hub_env.hub_root)
-    app = create_hub_app(hub_env.hub_root)
-    client = TestClient(app)
-
-    resp = client.get("/hub/pma/docs")
-    assert resp.status_code == 200
-    payload = resp.json()
-    assert "docs" in payload
-    docs = payload["docs"]
-    assert isinstance(docs, list)
-    doc_names = [doc["name"] for doc in docs]
-    assert doc_names == [
-        "AGENTS.md",
-        "active_context.md",
-        "context_log.md",
-        "ABOUT_CAR.md",
-        "prompt.md",
-    ]
-    for doc in docs:
-        assert "name" in doc
-        assert "exists" in doc
-        if doc["exists"]:
-            assert "size" in doc
-            assert "mtime" in doc
-        if doc["name"] == "active_context.md":
-            assert "line_count" in doc
-
-
-def test_pma_docs_list_includes_auto_prune_metadata(hub_env) -> None:
-    seed_hub_files(hub_env.hub_root, force=True)
-    _enable_pma(hub_env.hub_root)
-    active_context = (
-        hub_env.hub_root / ".codex-autorunner" / "pma" / "docs" / "active_context.md"
-    )
-    active_context.write_text(
-        "\n".join(f"line {idx}" for idx in range(220)), encoding="utf-8"
-    )
-    state = maybe_auto_prune_active_context(hub_env.hub_root, max_lines=50)
-    assert state is not None
-
-    app = create_hub_app(hub_env.hub_root)
-    client = TestClient(app)
-
-    resp = client.get("/hub/pma/docs")
-    assert resp.status_code == 200
-    payload = resp.json()
-    meta = payload.get("active_context_auto_prune")
-    assert isinstance(meta, dict)
-    assert meta.get("line_count_before") == 220
-    assert meta.get("line_budget") == 50
-    assert isinstance(meta.get("last_auto_pruned_at"), str)
-
-
-def test_pma_docs_get(hub_env) -> None:
-    seed_hub_files(hub_env.hub_root, force=True)
-    _enable_pma(hub_env.hub_root)
-    app = create_hub_app(hub_env.hub_root)
-    client = TestClient(app)
-
-    resp = client.get("/hub/pma/docs/AGENTS.md")
-    assert resp.status_code == 200
-    payload = resp.json()
-    assert payload["name"] == "AGENTS.md"
-    assert "content" in payload
-    assert isinstance(payload["content"], str)
-
-
-def test_pma_docs_get_nonexistent(hub_env) -> None:
-    seed_hub_files(hub_env.hub_root, force=True)
-    _enable_pma(hub_env.hub_root)
-    app = create_hub_app(hub_env.hub_root)
-    client = TestClient(app)
-
-    # Delete the canonical doc, then try to get it
-    pma_dir = hub_env.hub_root / ".codex-autorunner" / "pma"
-    docs_agents_path = pma_dir / "docs" / "AGENTS.md"
-    if docs_agents_path.exists():
-        docs_agents_path.unlink()
-
-    resp = client.get("/hub/pma/docs/AGENTS.md")
-    assert resp.status_code == 404
-    assert "not found" in resp.json()["detail"].lower()
-
-
-def test_pma_docs_list_migrates_legacy_doc_into_canonical(hub_env) -> None:
-    seed_hub_files(hub_env.hub_root, force=True)
-    _enable_pma(hub_env.hub_root)
-    app = create_hub_app(hub_env.hub_root)
-    client = TestClient(app)
-
-    pma_dir = hub_env.hub_root / ".codex-autorunner" / "pma"
-    docs_path = pma_dir / "docs" / "active_context.md"
-    legacy_path = pma_dir / "active_context.md"
-
-    docs_path.unlink(missing_ok=True)
-    legacy_content = "# Legacy copy\n\n- migrated\n"
-    legacy_path.write_text(legacy_content, encoding="utf-8")
-
-    resp = client.get("/hub/pma/docs")
-    assert resp.status_code == 200
-
-    assert docs_path.exists()
-    assert docs_path.read_text(encoding="utf-8") == legacy_content
-    assert not legacy_path.exists()
-
-    resp = client.get("/hub/pma/docs/active_context.md")
-    assert resp.status_code == 200
-    assert resp.json()["content"] == legacy_content
-
-
-def test_pma_docs_put(hub_env) -> None:
-    seed_hub_files(hub_env.hub_root, force=True)
-    _enable_pma(hub_env.hub_root)
-    app = create_hub_app(hub_env.hub_root)
-    client = TestClient(app)
-
-    new_content = "# AGENTS\n\nNew content"
-    resp = client.put("/hub/pma/docs/AGENTS.md", json={"content": new_content})
-    assert resp.status_code == 200
-    payload = resp.json()
-    assert payload["name"] == "AGENTS.md"
-    assert payload["status"] == "ok"
-
-    # Verify the content was saved
-    resp = client.get("/hub/pma/docs/AGENTS.md")
-    assert resp.status_code == 200
-    assert resp.json()["content"] == new_content
-
-
-def test_pma_docs_put_writes_via_to_thread(
-    hub_env, monkeypatch: pytest.MonkeyPatch
-) -> None:
-    seed_hub_files(hub_env.hub_root, force=True)
-    _enable_pma(hub_env.hub_root)
-    app = create_hub_app(hub_env.hub_root)
-    client = TestClient(app)
-    to_thread_calls: list[tuple[object, tuple[Any, ...], dict[str, Any]]] = []
-
-    async def _fake_to_thread(func, /, *args, **kwargs):
-        to_thread_calls.append((func, args, kwargs))
-        return func(*args, **kwargs)
-
-    monkeypatch.setattr(pma_routes.asyncio, "to_thread", _fake_to_thread)
-
-    resp = client.put("/hub/pma/docs/AGENTS.md", json={"content": "# AGENTS\n\nText"})
-    assert resp.status_code == 200
-    assert any(func is pma_routes.atomic_write for func, _, _ in to_thread_calls)
-
-
-def test_pma_docs_put_invalid_name(hub_env) -> None:
-    seed_hub_files(hub_env.hub_root, force=True)
-    _enable_pma(hub_env.hub_root)
-    app = create_hub_app(hub_env.hub_root)
-    client = TestClient(app)
-
-    resp = client.put("/hub/pma/docs/invalid.md", json={"content": "test"})
-    assert resp.status_code == 400
-    assert "Unknown doc name" in resp.json()["detail"]
-
-
-def test_pma_docs_put_too_large(hub_env) -> None:
-    seed_hub_files(hub_env.hub_root, force=True)
-    _enable_pma(hub_env.hub_root)
-    app = create_hub_app(hub_env.hub_root)
-    client = TestClient(app)
-
-    large_content = "x" * 500_001
-    resp = client.put("/hub/pma/docs/AGENTS.md", json={"content": large_content})
-    assert resp.status_code == 413
-    assert "too large" in resp.json()["detail"].lower()
-
-
-def test_pma_docs_put_invalid_content_type(hub_env) -> None:
-    seed_hub_files(hub_env.hub_root, force=True)
-    _enable_pma(hub_env.hub_root)
-    app = create_hub_app(hub_env.hub_root)
-    client = TestClient(app)
-
-    resp = client.put("/hub/pma/docs/AGENTS.md", json={"content": 123})
-    assert resp.status_code == 422
-    detail = resp.json()["detail"]
-    # FastAPI returns a list of validation errors
-    if isinstance(detail, list):
-        assert any("content" in str(err) for err in detail)
-    else:
-        assert "content" in str(detail)
-
-
-def test_pma_context_snapshot(hub_env) -> None:
-    seed_hub_files(hub_env.hub_root, force=True)
-    _enable_pma(hub_env.hub_root)
-    app = create_hub_app(hub_env.hub_root)
-    client = TestClient(app)
-
-    pma_dir = hub_env.hub_root / ".codex-autorunner" / "pma"
-    docs_dir = pma_dir / "docs"
-    docs_dir.mkdir(parents=True, exist_ok=True)
-    active_path = docs_dir / "active_context.md"
-    active_content = "# Active Context\n\n- alpha\n- beta\n"
-    active_path.write_text(active_content, encoding="utf-8")
-
-    resp = client.post("/hub/pma/context/snapshot", json={"reset": True})
-    assert resp.status_code == 200
-    payload = resp.json()
-    assert payload["status"] == "ok"
-    assert payload["active_context_line_count"] == len(active_content.splitlines())
-    assert payload["reset"] is True
-
-    log_content = (docs_dir / "context_log.md").read_text(encoding="utf-8")
-    assert "## Snapshot:" in log_content
-    assert active_content in log_content
-    assert active_path.read_text(encoding="utf-8") == pma_active_context_content()
-
-
-def test_pma_context_snapshot_writes_via_to_thread(
-    hub_env, monkeypatch: pytest.MonkeyPatch
-) -> None:
-    seed_hub_files(hub_env.hub_root, force=True)
-    _enable_pma(hub_env.hub_root)
-    app = create_hub_app(hub_env.hub_root)
-    client = TestClient(app)
-    to_thread_calls: list[tuple[object, tuple[Any, ...], dict[str, Any]]] = []
-
-    async def _fake_to_thread(func, /, *args, **kwargs):
-        to_thread_calls.append((func, args, kwargs))
-        return func(*args, **kwargs)
-
-    monkeypatch.setattr(pma_routes.asyncio, "to_thread", _fake_to_thread)
-
-    pma_dir = hub_env.hub_root / ".codex-autorunner" / "pma"
-    docs_dir = pma_dir / "docs"
-    docs_dir.mkdir(parents=True, exist_ok=True)
-    active_path = docs_dir / "active_context.md"
-    active_path.write_text("# Active Context\n\n- alpha\n", encoding="utf-8")
-
-    resp = client.post("/hub/pma/context/snapshot", json={"reset": True})
-    assert resp.status_code == 200
-    assert any(func is pma_routes.ensure_pma_docs for func, _, _ in to_thread_calls)
-    assert any(func is pma_routes.atomic_write for func, _, _ in to_thread_calls)
-
-
-def test_pma_docs_disabled(hub_env) -> None:
-    _disable_pma(hub_env.hub_root)
-    app = create_hub_app(hub_env.hub_root)
-    client = TestClient(app)
-
-    resp = client.get("/hub/pma/docs")
-    assert resp.status_code == 404
-
-    resp = client.get("/hub/pma/docs/AGENTS.md")
-    assert resp.status_code == 404
-
-
-def test_pma_automation_subscription_endpoints(hub_env) -> None:
-    _enable_pma(hub_env.hub_root)
-    app = create_hub_app(hub_env.hub_root)
-
-    class FakeAutomationStore:
-        def __init__(self) -> None:
-            self.created_payloads: list[dict[str, Any]] = []
-            self.list_filters: list[dict[str, Any]] = []
-            self.deleted_ids: list[str] = []
-
-        def create_subscription(self, payload: dict[str, Any]) -> dict[str, Any]:
-            self.created_payloads.append(dict(payload))
-            return {"subscription_id": "sub-1", **payload}
-
-        def list_subscriptions(self, **filters: Any) -> list[dict[str, Any]]:
-            self.list_filters.append(dict(filters))
-            return [{"subscription_id": "sub-1", "thread_id": "thread-1"}]
-
-        def cancel_subscription(self, subscription_id: str) -> dict[str, Any]:
-            self.deleted_ids.append(subscription_id)
-            return {"deleted": True}
-
-    fake_store = FakeAutomationStore()
-    app.state.hub_supervisor.get_pma_automation_store = lambda: fake_store
-
-    with TestClient(app) as client:
-        create_resp = client.post(
-            "/hub/pma/subscriptions",
-            json={
-                "thread_id": "thread-1",
-                "from_state": "running",
-                "to_state": "completed",
-                "reason": "manual",
-                "timestamp": "2026-03-01T12:00:00Z",
-            },
-        )
-        assert create_resp.status_code == 200
-        created = create_resp.json()["subscription"]
-        assert created["subscription_id"] == "sub-1"
-        assert created["thread_id"] == "thread-1"
-
-        list_resp = client.get(
-            "/hub/pma/automation/subscriptions",
-            params={"thread_id": "thread-1", "limit": 5},
-        )
-        assert list_resp.status_code == 200
-        listed = list_resp.json()["subscriptions"]
-        assert listed and listed[0]["subscription_id"] == "sub-1"
-
-        delete_resp = client.delete("/hub/pma/subscriptions/sub-1")
-        assert delete_resp.status_code == 200
-        assert delete_resp.json()["status"] == "ok"
-        assert delete_resp.json()["subscription_id"] == "sub-1"
-
-    assert fake_store.created_payloads
-    assert fake_store.created_payloads[0]["from_state"] == "running"
-    assert fake_store.created_payloads[0]["to_state"] == "completed"
-    assert (
-        fake_store.list_filters
-        and fake_store.list_filters[0]["thread_id"] == "thread-1"
-    )
-    assert fake_store.deleted_ids == ["sub-1"]
-
-
-@pytest.mark.parametrize(
-    ("request_body", "expected_event_types"),
-    [
-        (
-            {
-                "event_type": "managed_thread_completed",
-                "thread_id": "thread-1",
-            },
-            ["managed_thread_completed"],
-        ),
-        (
-            {
-                "event_types": [
-                    "managed_thread_completed",
-                    "managed_thread_failed",
-                ],
-                "thread_id": "thread-2",
-            },
-            ["managed_thread_completed", "managed_thread_failed"],
-        ),
-    ],
-)
-def test_pma_automation_subscription_create_normalizes_event_type_aliases(
-    hub_env,
-    request_body: dict[str, Any],
-    expected_event_types: list[str],
-) -> None:
-    _enable_pma(hub_env.hub_root)
-    app = create_hub_app(hub_env.hub_root)
-
-    class FakeAutomationStore:
-        def __init__(self) -> None:
-            self.created_payloads: list[dict[str, Any]] = []
-
-        def create_subscription(self, payload: dict[str, Any]) -> dict[str, Any]:
-            self.created_payloads.append(dict(payload))
-            return {"subscription_id": "sub-event-types-1", **payload}
-
-    fake_store = FakeAutomationStore()
-    app.state.hub_supervisor.get_pma_automation_store = lambda: fake_store
-
-    with TestClient(app) as client:
-        create_resp = client.post("/hub/pma/subscriptions", json=request_body)
-
-    assert create_resp.status_code == 200
-    assert fake_store.created_payloads == [
-        {
-            "thread_id": request_body["thread_id"],
-            "event_types": expected_event_types,
-        }
-    ]
-
-
-def test_pma_automation_subscription_create_accepts_nested_filter_payload(
-    hub_env,
-) -> None:
-    _enable_pma(hub_env.hub_root)
-    app = create_hub_app(hub_env.hub_root)
-
-    class FakeAutomationStore:
-        def __init__(self) -> None:
-            self.created_payloads: list[dict[str, Any]] = []
-
-        def create_subscription(self, payload: dict[str, Any]) -> dict[str, Any]:
-            self.created_payloads.append(dict(payload))
-            return {"subscription_id": "sub-filter-1", **payload}
-
-    fake_store = FakeAutomationStore()
-    app.state.hub_supervisor.get_pma_automation_store = lambda: fake_store
-
-    with TestClient(app) as client:
-        create_resp = client.post(
-            "/hub/pma/subscriptions",
-            json={
-                "event_types": ["managed_thread_completed"],
-                "filter": {"thread_id": "thread-filter-1", "repoId": "repo-filter-1"},
-            },
-        )
-
-    assert create_resp.status_code == 200
-    assert fake_store.created_payloads == [
-        {
-            "event_types": ["managed_thread_completed"],
-            "thread_id": "thread-filter-1",
-            "repo_id": "repo-filter-1",
-        }
-    ]
-
-
-def test_pma_automation_subscription_create_rejects_unknown_filter_keys(
-    hub_env,
-) -> None:
-    _enable_pma(hub_env.hub_root)
-    app = create_hub_app(hub_env.hub_root)
-
-    class FakeAutomationStore:
-        def create_subscription(self, payload: dict[str, Any]) -> dict[str, Any]:
-            return {"subscription_id": "should-not-be-created", **payload}
-
-    app.state.hub_supervisor.get_pma_automation_store = lambda: FakeAutomationStore()
-
-    with TestClient(app) as client:
-        create_resp = client.post(
-            "/hub/pma/subscriptions",
-            json={
-                "event_types": ["managed_thread_completed"],
-                "filter": {"thread_id": "thread-filter-1", "workspace_id": "ws-1"},
-            },
-        )
-
-    assert create_resp.status_code == 400
-    assert "Unsupported subscription filter keys" in create_resp.json()["detail"]
-    assert "workspace_id" in create_resp.json()["detail"]
-
-
-def test_pma_automation_subscription_create_rejects_conflicting_filter_values(
-    hub_env,
-) -> None:
-    _enable_pma(hub_env.hub_root)
-    app = create_hub_app(hub_env.hub_root)
-
-    class FakeAutomationStore:
-        def create_subscription(self, payload: dict[str, Any]) -> dict[str, Any]:
-            return {"subscription_id": "should-not-be-created", **payload}
-
-    app.state.hub_supervisor.get_pma_automation_store = lambda: FakeAutomationStore()
-
-    with TestClient(app) as client:
-        create_resp = client.post(
-            "/hub/pma/subscriptions",
-            json={
-                "event_types": ["managed_thread_completed"],
-                "thread_id": "thread-top-level",
-                "filter": {"thread_id": "thread-filter"},
-            },
-        )
-
-    assert create_resp.status_code == 400
-    assert "Conflicting values for thread_id" in create_resp.json()["detail"]
-
-
-def test_pma_automation_timer_endpoints(hub_env) -> None:
-    _enable_pma(hub_env.hub_root)
-    app = create_hub_app(hub_env.hub_root)
-
-    class FakeAutomationStore:
-        def __init__(self) -> None:
-            self.created_payloads: list[dict[str, Any]] = []
-            self.list_filters: list[dict[str, Any]] = []
-            self.touched: list[tuple[str, dict[str, Any]]] = []
-            self.cancelled: list[tuple[str, dict[str, Any]]] = []
-
-        def create_timer(self, payload: dict[str, Any]) -> dict[str, Any]:
-            self.created_payloads.append(dict(payload))
-            return {"timer_id": "timer-1", **payload}
-
-        def list_timers(self, **filters: Any) -> list[dict[str, Any]]:
-            self.list_filters.append(dict(filters))
-            return [{"timer_id": "timer-1", "thread_id": "thread-1"}]
-
-        def touch_timer(self, timer_id: str, payload: dict[str, Any]) -> dict[str, Any]:
-            self.touched.append((timer_id, dict(payload)))
-            return {"timer_id": timer_id, "touched": True}
-
-        def cancel_timer(
-            self, timer_id: str, payload: dict[str, Any]
-        ) -> dict[str, Any]:
-            self.cancelled.append((timer_id, dict(payload)))
-            return {"timer_id": timer_id, "cancelled": True}
-
-    fake_store = FakeAutomationStore()
-    app.state.hub_supervisor.get_pma_automation_store = lambda: fake_store
-
-    with TestClient(app) as client:
-        create_resp = client.post(
-            "/hub/pma/timers",
-            json={
-                "timer_type": "one_shot",
-                "delay_seconds": 1800,
-                "lane_id": "pma:lane-next",
-                "thread_id": "thread-1",
-                "from_state": "running",
-                "to_state": "failed",
-                "reason": "timeout",
-            },
-        )
-        assert create_resp.status_code == 200
-        created = create_resp.json()["timer"]
-        assert created["timer_id"] == "timer-1"
-        assert created["thread_id"] == "thread-1"
-
-        list_resp = client.get(
-            "/hub/pma/automation/timers",
-            params={"thread_id": "thread-1", "limit": 20},
-        )
-        assert list_resp.status_code == 200
-        listed = list_resp.json()["timers"]
-        assert listed and listed[0]["timer_id"] == "timer-1"
-
-        touch_resp = client.post(
-            "/hub/pma/timers/timer-1/touch",
-            json={"reason": "heartbeat"},
-        )
-        assert touch_resp.status_code == 200
-        assert touch_resp.json()["timer_id"] == "timer-1"
-
-        cancel_resp = client.post(
-            "/hub/pma/timers/timer-1/cancel",
-            json={"reason": "done"},
-        )
-        assert cancel_resp.status_code == 200
-        assert cancel_resp.json()["timer_id"] == "timer-1"
-
-    assert fake_store.created_payloads
-    assert fake_store.created_payloads[0]["timer_type"] == "one_shot"
-    assert fake_store.created_payloads[0]["delay_seconds"] == 1800
-    assert fake_store.created_payloads[0]["lane_id"] == "pma:lane-next"
-    assert fake_store.created_payloads[0]["to_state"] == "failed"
-    assert (
-        fake_store.list_filters
-        and fake_store.list_filters[0]["thread_id"] == "thread-1"
-    )
-    assert fake_store.touched == [("timer-1", {"reason": "heartbeat"})]
-    assert fake_store.cancelled == [("timer-1", {"reason": "done"})]
-
-
-def test_pma_automation_timer_create_accepts_nested_filter_payload(hub_env) -> None:
-    _enable_pma(hub_env.hub_root)
-    app = create_hub_app(hub_env.hub_root)
-
-    class FakeAutomationStore:
-        def __init__(self) -> None:
-            self.created_payloads: list[dict[str, Any]] = []
-
-        def create_timer(self, payload: dict[str, Any]) -> dict[str, Any]:
-            self.created_payloads.append(dict(payload))
-            return {"timer_id": "timer-filter-1", **payload}
-
-    fake_store = FakeAutomationStore()
-    app.state.hub_supervisor.get_pma_automation_store = lambda: fake_store
-
-    with TestClient(app) as client:
-        create_resp = client.post(
-            "/hub/pma/timers",
-            json={
-                "timer_type": "one_shot",
-                "delay_seconds": 30,
-                "filter": {"threadId": "thread-filter-1", "repo_id": "repo-filter-1"},
-            },
-        )
-
-    assert create_resp.status_code == 200
-    assert fake_store.created_payloads == [
-        {
-            "timer_type": "one_shot",
-            "delay_seconds": 30,
-            "thread_id": "thread-filter-1",
-            "repo_id": "repo-filter-1",
-        }
-    ]
-
-
-def test_pma_automation_timer_create_rejects_unknown_filter_keys(hub_env) -> None:
-    _enable_pma(hub_env.hub_root)
-    app = create_hub_app(hub_env.hub_root)
-
-    class FakeAutomationStore:
-        def create_timer(self, payload: dict[str, Any]) -> dict[str, Any]:
-            return {"timer_id": "should-not-be-created", **payload}
-
-    app.state.hub_supervisor.get_pma_automation_store = lambda: FakeAutomationStore()
-
-    with TestClient(app) as client:
-        create_resp = client.post(
-            "/hub/pma/timers",
-            json={
-                "timer_type": "one_shot",
-                "delay_seconds": 30,
-                "filter": {"thread_id": "thread-filter-1", "workspace_id": "ws-1"},
-            },
-        )
-
-    assert create_resp.status_code == 400
-    assert "Unsupported timer filter keys" in create_resp.json()["detail"]
-    assert "workspace_id" in create_resp.json()["detail"]
-
-
-def test_pma_automation_timer_create_rejects_conflicting_filter_values(
-    hub_env,
-) -> None:
-    _enable_pma(hub_env.hub_root)
-    app = create_hub_app(hub_env.hub_root)
-
-    class FakeAutomationStore:
-        def create_timer(self, payload: dict[str, Any]) -> dict[str, Any]:
-            return {"timer_id": "should-not-be-created", **payload}
-
-    app.state.hub_supervisor.get_pma_automation_store = lambda: FakeAutomationStore()
-
-    with TestClient(app) as client:
-        create_resp = client.post(
-            "/hub/pma/timers",
-            json={
-                "timer_type": "one_shot",
-                "delay_seconds": 30,
-                "thread_id": "thread-top-level",
-                "filter": {"thread_id": "thread-filter"},
-            },
-        )
-
-    assert create_resp.status_code == 400
-    assert "Conflicting values for thread_id" in create_resp.json()["detail"]
-
-
-def test_pma_automation_timer_touch_rejects_unknown_keys(hub_env) -> None:
-    _enable_pma(hub_env.hub_root)
-    app = create_hub_app(hub_env.hub_root)
-
-    class FakeAutomationStore:
-        def touch_timer(self, timer_id: str, payload: dict[str, Any]) -> dict[str, Any]:
-            return {"timer_id": timer_id, "payload": dict(payload)}
-
-    app.state.hub_supervisor.get_pma_automation_store = lambda: FakeAutomationStore()
-
-    with TestClient(app) as client:
-        response = client.post(
-            "/hub/pma/timers/timer-1/touch",
-            json={"reason": "heartbeat", "oops": True},
-        )
-
-    assert response.status_code == 400
-    assert "Unsupported timer touch keys" in response.json()["detail"]
-
-
-def test_pma_automation_timer_cancel_rejects_unknown_keys(hub_env) -> None:
-    _enable_pma(hub_env.hub_root)
-    app = create_hub_app(hub_env.hub_root)
-
-    class FakeAutomationStore:
-        def cancel_timer(
-            self, timer_id: str, payload: dict[str, Any]
-        ) -> dict[str, Any]:
-            return {"timer_id": timer_id, "payload": dict(payload)}
-
-    app.state.hub_supervisor.get_pma_automation_store = lambda: FakeAutomationStore()
-
-    with TestClient(app) as client:
-        response = client.post(
-            "/hub/pma/timers/timer-1/cancel",
-            json={"reason": "done", "unexpected": "value"},
-        )
-
-    assert response.status_code == 400
-    assert "Unsupported timer cancel keys" in response.json()["detail"]
-
-
-def test_pma_automation_subscription_alias_endpoint_supports_kwargs_only_store(
-    hub_env,
-) -> None:
-    _enable_pma(hub_env.hub_root)
-    app = create_hub_app(hub_env.hub_root)
-
-    class FakeAutomationStore:
-        def __init__(self) -> None:
-            self.create_calls: list[dict[str, Any]] = []
-            self.deleted_ids: list[str] = []
-
-        def create_subscription(
-            self, *, thread_id: Optional[str] = None, lane_id: Optional[str] = None
-        ) -> dict[str, Any]:
-            self.create_calls.append({"thread_id": thread_id, "lane_id": lane_id})
-            return {
-                "subscription_id": "sub-alias-1",
-                "thread_id": thread_id,
-                "lane_id": lane_id,
-            }
-
-        def cancel_subscription(self, subscription_id: str) -> bool:
-            self.deleted_ids.append(subscription_id)
-            return True
-
-    fake_store = FakeAutomationStore()
-    app.state.hub_supervisor.get_pma_automation_store = lambda: fake_store
-
-    with TestClient(app) as client:
-        create_resp = client.post(
-            "/hub/pma/automation/subscriptions",
-            json={"thread_id": "thread-alias", "lane_id": "pma:lane-alias"},
-        )
-        assert create_resp.status_code == 200
-        subscription = create_resp.json()["subscription"]
-        assert subscription["subscription_id"] == "sub-alias-1"
-        assert subscription["thread_id"] == "thread-alias"
-        assert subscription["lane_id"] == "pma:lane-alias"
-
-        delete_resp = client.delete("/hub/pma/automation/subscriptions/sub-alias-1")
-        assert delete_resp.status_code == 200
-        delete_payload = delete_resp.json()
-        assert delete_payload["status"] == "ok"
-        assert delete_payload["subscription_id"] == "sub-alias-1"
-
-    assert fake_store.create_calls == [
-        {"thread_id": "thread-alias", "lane_id": "pma:lane-alias"}
-    ]
-    assert fake_store.deleted_ids == ["sub-alias-1"]
-
-
-def test_pma_automation_timer_alias_endpoint_supports_fallback_method_signatures(
-    hub_env,
-) -> None:
-    _enable_pma(hub_env.hub_root)
-    app = create_hub_app(hub_env.hub_root)
-
-    class FakeAutomationStore:
-        def __init__(self) -> None:
-            self.create_calls: list[dict[str, Any]] = []
-            self.cancelled_ids: list[str] = []
-
-        def create_timer(
-            self,
-            *,
-            timer_type: Optional[str] = None,
-            idle_seconds: Optional[int] = None,
-        ) -> dict[str, Any]:
-            self.create_calls.append(
-                {"timer_type": timer_type, "idle_seconds": idle_seconds}
-            )
-            return {
-                "timer_id": "timer-alias-1",
-                "timer_type": timer_type,
-                "idle_seconds": idle_seconds,
-            }
-
-        def cancel_timer(self, timer_id: str) -> bool:
-            self.cancelled_ids.append(timer_id)
-            return True
-
-    fake_store = FakeAutomationStore()
-    app.state.hub_supervisor.get_pma_automation_store = lambda: fake_store
-
-    with TestClient(app) as client:
-        create_resp = client.post(
-            "/hub/pma/automation/timers",
-            json={"timer_type": "watchdog", "idle_seconds": 45},
-        )
-        assert create_resp.status_code == 200
-        timer = create_resp.json()["timer"]
-        assert timer["timer_id"] == "timer-alias-1"
-        assert timer["timer_type"] == "watchdog"
-        assert timer["idle_seconds"] == 45
-
-        cancel_resp = client.delete("/hub/pma/automation/timers/timer-alias-1")
-        assert cancel_resp.status_code == 200
-        cancel_payload = cancel_resp.json()
-        assert cancel_payload["status"] == "ok"
-        assert cancel_payload["timer_id"] == "timer-alias-1"
-
-    assert fake_store.create_calls == [{"timer_type": "watchdog", "idle_seconds": 45}]
-    assert fake_store.cancelled_ids == ["timer-alias-1"]
-
-
-def test_pma_automation_watchdog_timer_create(hub_env) -> None:
-    _enable_pma(hub_env.hub_root)
-    app = create_hub_app(hub_env.hub_root)
-
-    class FakeAutomationStore:
-        def __init__(self) -> None:
-            self.created_payloads: list[dict[str, Any]] = []
-
-        def create_timer(self, payload: dict[str, Any]) -> dict[str, Any]:
-            self.created_payloads.append(dict(payload))
-            return {"timer_id": "watchdog-1", **payload}
-
-    fake_store = FakeAutomationStore()
-    app.state.hub_supervisor.get_pma_automation_store = lambda: fake_store
-
-    with TestClient(app) as client:
-        create_resp = client.post(
-            "/hub/pma/timers",
-            json={
-                "timer_type": "watchdog",
-                "idle_seconds": 300,
-                "thread_id": "thread-1",
-                "reason": "watchdog_stalled",
-            },
-        )
-        assert create_resp.status_code == 200
-        payload = create_resp.json()["timer"]
-        assert payload["timer_id"] == "watchdog-1"
-
-    assert fake_store.created_payloads
-    assert fake_store.created_payloads[0]["timer_type"] == "watchdog"
-    assert fake_store.created_payloads[0]["idle_seconds"] == 300
-
-
-def test_pma_automation_timer_rejects_invalid_due_at(hub_env) -> None:
-    _enable_pma(hub_env.hub_root)
-    app = create_hub_app(hub_env.hub_root)
-
-    class FakeAutomationStore:
-        def __init__(self) -> None:
-            self.created_payloads: list[dict[str, Any]] = []
-
-        def create_timer(self, payload: dict[str, Any]) -> dict[str, Any]:
-            self.created_payloads.append(dict(payload))
-            return {"timer_id": "timer-1", **payload}
-
-    fake_store = FakeAutomationStore()
-    app.state.hub_supervisor.get_pma_automation_store = lambda: fake_store
-
-    with TestClient(app) as client:
-        response = client.post(
-            "/hub/pma/timers",
-            json={"timer_type": "one_shot", "due_at": "not-a-timestamp"},
-        )
-        assert response.status_code == 422
-
-    assert fake_store.created_payloads == []
-
-
-def test_pma_automation_timer_rejects_unknown_subscription_id(hub_env) -> None:
-    _enable_pma(hub_env.hub_root)
-    app = create_hub_app(hub_env.hub_root)
-
-    with TestClient(app) as client:
-        response = client.post(
-            "/hub/pma/timers",
-            json={
-                "timer_type": "one_shot",
-                "delay_seconds": 60,
-                "subscription_id": "missing-sub",
-            },
-        )
-        assert response.status_code == 400
-        assert response.json()["detail"] == "Unknown subscription_id: missing-sub"
-
-
-def test_pma_orchestration_service_integration_for_thread_operations(
-    hub_env, monkeypatch: pytest.MonkeyPatch
-) -> None:
-    _enable_pma(hub_env.hub_root)
-
-    class FakeService:
-        def __init__(self) -> None:
-            self.calls: list[str] = []
-
-        def get_thread_target(self, thread_target_id: str):
-            self.calls.append(f"get_thread_target:{thread_target_id}")
-            return None
-
-        def get_running_execution(self, thread_target_id: str):
-            self.calls.append(f"get_running_execution:{thread_target_id}")
-            return None
-
-        def get_latest_execution(self, thread_target_id: str):
-            self.calls.append(f"get_latest_execution:{thread_target_id}")
-            return None
-
-    fake_service = FakeService()
-    monkeypatch.setattr(
-        tail_stream,
-        "build_managed_thread_orchestration_service",
-        lambda request: fake_service,
-    )
-    app = create_hub_app(hub_env.hub_root)
-
-    client = TestClient(app)
-
-    client.get("/hub/pma/threads/thread-1/status")
-    assert any(
-        call.startswith("get_thread_target:thread-1") for call in fake_service.calls
-    )
-
-    fake_service.calls.clear()
-    client.get("/hub/pma/threads/thread-1/tail")
-    assert any(
-        call.startswith("get_thread_target:thread-1") for call in fake_service.calls
-    )
 
 
 def test_pma_active_status_distinguishes_running_from_idle(hub_env) -> None:
